+++ cj2pgsql/resources/post_import.sql
@@ -5,6 +5,7 @@ create index if not exists import_meta_source_file_idx on {schema}.import_meta u
 -- cj_object indexes
 create index if not exists cj_object_type_idx on {schema}.cj_object using btree("type");
 create index if not exists cj_object_ground_gix on {schema}.cj_object using gist(ground_geometry);
+create index lod on {schema}.cj_object using gin (geometry);
 
 -- family indexes
 create index if not exists family_parent_idx on {schema}.family using btree(parent_id);