+++ changelog.md
@@ -21,6 +21,7 @@ for soon-to-be removed features.
 `Added`
 - user can choose between an ordinary (-x) and partial index (-px) on any attribute
 - lod is indexed using GIN by default
+- use multi inserts for performance
 
 
 ## [0.0.8a] - 2022-10-31